Passage de certaine fonction en const & (supression warnings) Correction methode qHash de SxRefPtr pour 64bit (supression warning) Utilisation des conteneur QT Correction comparaison signed/unsigned (supression warning) Correction troncation de pointeur (supression warning)

+++ Sources/SwCore/Types/SwDouble.cpp
@@ -101,7 +101,7 @@ int SwDouble::getPrecision()
 }
 
 //---------------------------------------------------------------------------------
-void SwDouble::setPrecision( double prec)
+void SwDouble::setPrecision( int prec)
 {
 	_precision = prec;
 	_doubleValidator->setDecimals(prec);
